+++ CHANGES.md
@@ -6,6 +6,7 @@ Apollo Java 2.4.0
 
 ------------------
 * [Fix the Cannot enhance @Configuration bean definition issue](https://github.com/apolloconfig/apollo-java/pull/82)
+* [Feature openapi query namespace support not fill item](https://github.com/apolloconfig/apollo-java/pull/83)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo-java/milestone/4?closed=1)